# Conflicts:
#	.idea/misc.xml
#	client/src/main/java/gov/iti/jets/client/callBack/CheckConnection.java
#	server/src/main/java/gov/iti/jets/server/controller/ConnectionController.java

+++ common/src/main/java/gov/iti/jets/common/network/client/IClient.java
@@ -7,5 +7,5 @@ import java.util.LinkedList;
 
 public interface IClient extends Remote {
     public void receive() throws RemoteException;
-    public void receiveMessage(int chatId, String senderId, String message) throws RemoteException;
+    public void receiveMessage(long chatId, String senderId, String message) throws RemoteException;
 }
